일부 항목 수정

+++ src/api/book.api.js
@@ -3,19 +3,19 @@ import { apiClient, fileApiClient } from "./client";
 /* 관리자 요청 api */
 /* DB 내 도서 등록 api */
 export const registerBookByAdmin = async (formData) => {
-    const res = await fileApiClient.post('/book/register', formData);
+    const res = await fileApiClient.post('/book/admin/register', formData);
     return res.data;
 }
 
 /*도서 정보 수정 api*/
 export const updateBookByBookId = async (bookId, data) => {
-    const res = await fileApiClient.patch(`/book/${bookId}`, data);
+    const res = await fileApiClient.patch(`/book/admin/${bookId}`, data);
     return res.data;
 }
 
 /* 도서 삭제 */
 export const deleteBooks = async (data) => {
-    const res = await apiClient.delete('/book', { data });
+    const res = await apiClient.delete('/book/admin', { data });
     return res.data;
 }
 
@@ -34,7 +34,7 @@ export const getAllBookList = async (data) => {
 
 /* 카카오 + 국립 중앙 도서관 api를 활용한 도서 상세 정보 검색 api */
 export const searchBooksWithAPi = async (query) => {
-    const res = await apiClient.get('/book/library/search', {
+    const res = await apiClient.get('/book/admin/library/search', {
         params: { query }
     });
     return res.data;
@@ -47,7 +47,7 @@ export const searchBookWithAPiByIsbn = async (isbn) => {
 
 /* 도서 중복 체크 api */
 export const checkBookDuplicate = async (isbn) => {
-    const res = await apiClient.get(`/book/is-duplicate-isbn?isbn=${isbn}`);
+    const res = await apiClient.get(`/book/admin/is-duplicate-isbn?isbn=${isbn}`);
     return res.data;
 };
